feat: change screen appearance order
