+++ app/index.tsx
@@ -1,81 +1,44 @@
 import React, { useState } from 'react';
 import { View, TextInput, StyleSheet, TouchableOpacity, Text, Alert, Image } from 'react-native';
-import { addDoc, collection, query, where, getDocs } from 'firebase/firestore';
+import { getDocs, query, collection, where } from 'firebase/firestore';
 import { database } from '@/config/firebaseConfig';
 import { useRouter } from 'expo-router';
-import { Snackbar } from 'react-native-paper';
 import { FontAwesome } from '@expo/vector-icons';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import Logo from '@/assets/images/Logo.svg';
 
-export default function RegisterScreen() {
-  const [name, setName] = useState('');
+export default function LoginScreen() {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-  const [visible, setVisible] = useState(false);
-  const [error, setError] = useState('');
-
   const router = useRouter();
 
-  async function handleRegister() {
-    if (!name || !email || !password || !confirmPassword) {
-      setError('Por favor, preencha todos os campos');
-      return;
-    }
-
-    if (password !== confirmPassword) {
-      setError('As senhas não coincidem');
-      setTimeout(() => setError(''), 4000);
-      return;
-    }
-
-    const allowedDomains = ['gmail.com', 'hotmail.com'];
-    const emailDomain = email.split('@')[1];
-
-    if (!allowedDomains.includes(emailDomain)) {
-      setError('O email deve ser do domínio gmail.com ou hotmail.com');
-      setTimeout(() => setError(''), 4000);
-      return;
-    }
-
-    // Verificar se o email já está em uso
-    const usersRef = collection(database, 'users');
-    const q = query(usersRef, where('email', '==', email));
-    const querySnapshot = await getDocs(q);
-
-    if (!querySnapshot.empty) {
-      setError('Este email já está sendo usado por outro usuário');
-      setTimeout(() => setError(''), 2000);
-      return;
-    }
-
+  async function handleLogin() {
     try {
-      await addDoc(usersRef, {
-        name,
-        email,
-        password,
-      });
+      const q = query(collection(database, 'users'), where('email', '==', email), where('password', '==', password));
+      const querySnapshot = await getDocs(q);
 
-      // Fetch the newly created user data
-      const newUserQuery = query(usersRef, where('email', '==', email));
-      const newUserSnapshot = await getDocs(newUserQuery);
-      await AsyncStorage.setItem('userEmail', email);
+      if (!querySnapshot.empty) {
+        const userData = querySnapshot.docs[0].data();
+        
+        // Save the user email to AsyncStorage
+        await AsyncStorage.setItem('userEmail', userData.email);
 
-      if (!newUserSnapshot.empty) {
-        const newUser = newUserSnapshot.docs[0].data();
         router.push({
-          pathname: '/onBoard',
-          params: newUser,
+          pathname: '/(tabs)/HomeScreen',
+          params: {
+            email: userData.email,
+            name: userData.name,
+            phone: userData.phone,
+          },
         });
-      }
-
-      setVisible(true);
-    } catch (e: unknown) {
-      if (e instanceof Error) {
-        alert('Erro ao adicionar usuário: ' + e.message);
       } else {
-        alert('Erro desconhecido ao adicionar usuário');
+        Alert.alert('Usuário não encontrado ou senha incorreta');
+      }
+    } catch (error) {
+      if (error instanceof Error) {
+        Alert.alert('Erro ao fazer login: ' + error.message);
+      } else {
+        Alert.alert('Erro desconhecido ao fazer login');
       }
     }
   }
@@ -83,17 +46,7 @@ export default function RegisterScreen() {
   return (
     <View style={styles.container}>
       <Logo/>
-      <Text style={styles.subtitle}>Cadastro</Text>
-      <View style={styles.inputContainer}>
-        <FontAwesome name="user" size={24} color="black" />
-        <TextInput
-          style={styles.input}
-          placeholder="Nome"
-          placeholderTextColor="#aaa"
-          value={name}
-          onChangeText={setName}
-        />
-      </View>
+      <Text style={styles.subtitle}>Login</Text>
       <View style={styles.inputContainer}>
         <FontAwesome name="envelope" size={24} color="black" />
         <TextInput
@@ -103,6 +56,7 @@ export default function RegisterScreen() {
           value={email}
           onChangeText={setEmail}
           keyboardType="email-address"
+          autoCapitalize="none"
         />
       </View>
       <View style={styles.inputContainer}>
@@ -113,28 +67,16 @@ export default function RegisterScreen() {
           placeholderTextColor="#aaa"
           value={password}
           onChangeText={setPassword}
-          secureTextEntry={true}
+          secureTextEntry
         />
       </View>
-      <View style={styles.inputContainer}>
-        <FontAwesome name="lock" size={24} color="black" />
-        <TextInput
-          style={styles.input}
-          placeholder="Confirmar Senha"
-          placeholderTextColor="#aaa"
-          value={confirmPassword}
-          onChangeText={setConfirmPassword}
-          secureTextEntry={true}
-        />
-      </View>
-      {error ? <Text style={styles.errorText}>{error}</Text> : null}
-      <TouchableOpacity style={styles.button} onPress={handleRegister}>
-        <Text style={styles.buttonText}>Cadastrar</Text>
+      <TouchableOpacity style={styles.button} onPress={handleLogin}>
+        <Text style={styles.buttonText}>Login</Text>
       </TouchableOpacity>
-      <Text style={styles.loginText}>
-        Já possui uma conta?{' '}
-        <Text style={styles.loginLink} onPress={() => router.push('/loginscreen')}>
-          Faça login
+      <Text style={styles.registerText}>
+        Não possui conta?{' '}
+        <Text style={styles.registerLink} onPress={() => router.push('/register')}>
+          Faça seu cadastro
         </Text>
       </Text>
     </View>
@@ -196,19 +138,12 @@ const styles = StyleSheet.create({
     color: '#fff',
     fontSize: 18,
   },
-  loginText: {
+  registerText: {
     color: '#000',
     fontSize: 14,
   },
-  loginLink: {
+  registerLink: {
     color: '#FFA500',
     fontWeight: 'bold',
-  },
-  snackbar: {
-    backgroundColor: 'green',
-  },
-  errorText: {
-    color: 'red',
-    marginBottom: 10,
   },
 });
