--- app/loginscreen/index.tsx
@@ -1,149 +1,0 @@
-import React, { useState } from 'react';
-import { View, TextInput, StyleSheet, TouchableOpacity, Text, Alert, Image } from 'react-native';
-import { getDocs, query, collection, where } from 'firebase/firestore';
-import { database } from '@/config/firebaseConfig';
-import { useRouter } from 'expo-router';
-import { FontAwesome } from '@expo/vector-icons';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import Logo from '@/assets/images/Logo.svg';
-
-export default function LoginScreen() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const router = useRouter();
-
-  async function handleLogin() {
-    try {
-      const q = query(collection(database, 'users'), where('email', '==', email), where('password', '==', password));
-      const querySnapshot = await getDocs(q);
-
-      if (!querySnapshot.empty) {
-        const userData = querySnapshot.docs[0].data();
-        
-        // Save the user email to AsyncStorage
-        await AsyncStorage.setItem('userEmail', userData.email);
-
-        router.push({
-          pathname: '/(tabs)/HomeScreen',
-          params: {
-            email: userData.email,
-            name: userData.name,
-            phone: userData.phone,
-          },
-        });
-      } else {
-        Alert.alert('Usuário não encontrado ou senha incorreta');
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        Alert.alert('Erro ao fazer login: ' + error.message);
-      } else {
-        Alert.alert('Erro desconhecido ao fazer login');
-      }
-    }
-  }
-
-  return (
-    <View style={styles.container}>
-      <Logo/>
-      <Text style={styles.subtitle}>Login</Text>
-      <View style={styles.inputContainer}>
-        <FontAwesome name="envelope" size={24} color="black" />
-        <TextInput
-          style={styles.input}
-          placeholder="Email"
-          placeholderTextColor="#aaa"
-          value={email}
-          onChangeText={setEmail}
-          keyboardType="email-address"
-          autoCapitalize="none"
-        />
-      </View>
-      <View style={styles.inputContainer}>
-        <FontAwesome name="lock" size={24} color="black" />
-        <TextInput
-          style={styles.input}
-          placeholder="Senha"
-          placeholderTextColor="#aaa"
-          value={password}
-          onChangeText={setPassword}
-          secureTextEntry
-        />
-      </View>
-      <TouchableOpacity style={styles.button} onPress={handleLogin}>
-        <Text style={styles.buttonText}>Login</Text>
-      </TouchableOpacity>
-      <Text style={styles.registerText}>
-        Não possui conta?{' '}
-        <Text style={styles.registerLink} onPress={() => router.push('/')}>
-          Faça seu cadastro
-        </Text>
-      </Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 16,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#f5f5f5',
-  },
-  logo: {
-    width: 210, 
-    height: 270, 
-    marginBottom: 30, 
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#FFA500',
-    marginVertical: 20,
-  },
-  subtitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#000',
-    marginBottom: 20,
-  },
-  inputContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderColor: '#ccc',
-    borderWidth: 1,
-    borderRadius: 5,
-    marginBottom: 12,
-    paddingHorizontal: 8,
-    backgroundColor: '#fff',
-    width: '80%',
-    height: 40,
-  },
-  input: {
-    flex: 1,
-    paddingLeft: 10,
-    color: '#000',
-  },
-  button: {
-    backgroundColor: '#FFA500',
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    borderRadius: 5,
-    marginVertical: 10,
-    width: '80%',
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: '#fff',
-    fontSize: 18,
-  },
-  registerText: {
-    color: '#000',
-    fontSize: 14,
-  },
-  registerLink: {
-    color: '#FFA500',
-    fontWeight: 'bold',
-  },
-});
