+++ app/register.tsx
@@ -1,0 +1,214 @@
+import React, { useState } from 'react';
+import { View, TextInput, StyleSheet, TouchableOpacity, Text, Alert, Image } from 'react-native';
+import { addDoc, collection, query, where, getDocs } from 'firebase/firestore';
+import { database } from '@/config/firebaseConfig';
+import { useRouter } from 'expo-router';
+import { Snackbar } from 'react-native-paper';
+import { FontAwesome } from '@expo/vector-icons';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import Logo from '@/assets/images/Logo.svg';
+
+export default function RegisterScreen() {
+  const [name, setName] = useState('');
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [visible, setVisible] = useState(false);
+  const [error, setError] = useState('');
+
+  const router = useRouter();
+
+  async function handleRegister() {
+    if (!name || !email || !password || !confirmPassword) {
+      setError('Por favor, preencha todos os campos');
+      return;
+    }
+
+    if (password !== confirmPassword) {
+      setError('As senhas não coincidem');
+      setTimeout(() => setError(''), 4000);
+      return;
+    }
+
+    const allowedDomains = ['gmail.com', 'hotmail.com'];
+    const emailDomain = email.split('@')[1];
+
+    if (!allowedDomains.includes(emailDomain)) {
+      setError('O email deve ser do domínio gmail.com ou hotmail.com');
+      setTimeout(() => setError(''), 4000);
+      return;
+    }
+
+    // Verificar se o email já está em uso
+    const usersRef = collection(database, 'users');
+    const q = query(usersRef, where('email', '==', email));
+    const querySnapshot = await getDocs(q);
+
+    if (!querySnapshot.empty) {
+      setError('Este email já está sendo usado por outro usuário');
+      setTimeout(() => setError(''), 2000);
+      return;
+    }
+
+    try {
+      await addDoc(usersRef, {
+        name,
+        email,
+        password,
+      });
+
+      // Fetch the newly created user data
+      const newUserQuery = query(usersRef, where('email', '==', email));
+      const newUserSnapshot = await getDocs(newUserQuery);
+      await AsyncStorage.setItem('userEmail', email);
+
+      if (!newUserSnapshot.empty) {
+        const newUser = newUserSnapshot.docs[0].data();
+        router.push({
+          pathname: '/onBoard',
+          params: newUser,
+        });
+      }
+
+      setVisible(true);
+    } catch (e: unknown) {
+      if (e instanceof Error) {
+        alert('Erro ao adicionar usuário: ' + e.message);
+      } else {
+        alert('Erro desconhecido ao adicionar usuário');
+      }
+    }
+  }
+
+  return (
+    <View style={styles.container}>
+      <Logo/>
+      <Text style={styles.subtitle}>Cadastro</Text>
+      <View style={styles.inputContainer}>
+        <FontAwesome name="user" size={24} color="black" />
+        <TextInput
+          style={styles.input}
+          placeholder="Nome"
+          placeholderTextColor="#aaa"
+          value={name}
+          onChangeText={setName}
+        />
+      </View>
+      <View style={styles.inputContainer}>
+        <FontAwesome name="envelope" size={24} color="black" />
+        <TextInput
+          style={styles.input}
+          placeholder="Email"
+          placeholderTextColor="#aaa"
+          value={email}
+          onChangeText={setEmail}
+          keyboardType="email-address"
+        />
+      </View>
+      <View style={styles.inputContainer}>
+        <FontAwesome name="lock" size={24} color="black" />
+        <TextInput
+          style={styles.input}
+          placeholder="Senha"
+          placeholderTextColor="#aaa"
+          value={password}
+          onChangeText={setPassword}
+          secureTextEntry={true}
+        />
+      </View>
+      <View style={styles.inputContainer}>
+        <FontAwesome name="lock" size={24} color="black" />
+        <TextInput
+          style={styles.input}
+          placeholder="Confirmar Senha"
+          placeholderTextColor="#aaa"
+          value={confirmPassword}
+          onChangeText={setConfirmPassword}
+          secureTextEntry={true}
+        />
+      </View>
+      {error ? <Text style={styles.errorText}>{error}</Text> : null}
+      <TouchableOpacity style={styles.button} onPress={handleRegister}>
+        <Text style={styles.buttonText}>Cadastrar</Text>
+      </TouchableOpacity>
+      <Text style={styles.loginText}>
+        Já possui uma conta?{' '}
+        <Text style={styles.loginLink} onPress={() => router.push('/')}>
+          Faça login
+        </Text>
+      </Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#f5f5f5',
+  },
+  logo: {
+    width: 210, 
+    height: 270, 
+    marginBottom: 30, 
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    color: '#FFA500',
+    marginVertical: 20,
+  },
+  subtitle: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: '#000',
+    marginBottom: 20,
+  },
+  inputContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    borderColor: '#ccc',
+    borderWidth: 1,
+    borderRadius: 5,
+    marginBottom: 12,
+    paddingHorizontal: 8,
+    backgroundColor: '#fff',
+    width: '80%',
+    height: 40,
+  },
+  input: {
+    flex: 1,
+    paddingLeft: 10,
+    color: '#000',
+  },
+  button: {
+    backgroundColor: '#FFA500',
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    borderRadius: 5,
+    marginVertical: 10,
+    width: '80%',
+    alignItems: 'center',
+  },
+  buttonText: {
+    color: '#fff',
+    fontSize: 18,
+  },
+  loginText: {
+    color: '#000',
+    fontSize: 14,
+  },
+  loginLink: {
+    color: '#FFA500',
+    fontWeight: 'bold',
+  },
+  snackbar: {
+    backgroundColor: 'green',
+  },
+  errorText: {
+    color: 'red',
+    marginBottom: 10,
+  },
+});
